requirements.txt
future
emoji
lxml
wget
gtts
beautifulsoup4
requests==2.20.0
python-telegram-bot==12.8
sqlalchemy==1.4.11
PyDictionary==2.0.1
psycopg2-binary
feedparser
hachoir
odmantic
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
jikanpy
speedtest-cli
coffeehouse==2.2.4
regex
bleach
markdown2>=2.4.0
wikipedia
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
TgCrypto
PyroGram
youtube-dl
youtube_search
asyncio
dateparser==1.0.0
pymongo==3.12.1
motor
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
bing_image_downloader
search_engine_parser
pytz
Skem
tswift
redis
aiofiles
pyromod
TgCrypto
googletrans==4.0.0-rc1
google-trans-new==1.1.9
gpytranslate
PyYAML>5.0
envparse
cloudscraper
pykeyboard
youtube-search
youtube-search-python
opencv-python-headless
img2pdf
multicolorcaptcha
youtube-search
phlogo
